pyproject.toml
[project]
name = "agenticapproach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-framework>=1.0.1",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.25.3",
    "azure-search-documents>=11.7.0b2",
    "fastapi>=0.133.0",
    "openai>=2.31.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.41.0",
]
